一旦コミット。axiosによるAPI呼び出し

+++ src/database/migrations/2023_09_19_221728_create_medicines_table.php
@@ -17,8 +17,8 @@ return new class extends Migration
             $table->foreignId('unit_id')->constrained()->onDelete('cascade');
             $table->string('name');
             $table->date('start_date');
-            $table->integer('dose_amount');
-            $table->integer('stock_amount');
+            $table->decimal('dose_amount', 8, 2); // 少数第二位まで
+            $table->decimal('stock_amount', 8, 2); // 少数第二位まで
             $table->timestamps();
         });
     }
